+++ EWSContacts/Module/functions/contacts/Get-GALContacts.ps1
@@ -1,7 +1,7 @@
 function Get-GALContacts {
 	<#
 	.SYNOPSIS
-		Uses Office 365 services to generate a list of contacts 
+		Uses Office 365 services to generate a list of contacts. Only includes contacts with an email address.
 	
 	.PARAMETER ConnectionUri
 		Used to connect to Office 365, by default this is https://outlook.office365.com/powershell-liveid/.
@@ -9,11 +9,14 @@ function Get-GALContacts {
 	.PARAMETER Credentials
 		Office 365 Admin Credentials
 
-	.PARAMETER RequirePhoneNumber
+	.PARAMETER ExcludeContactsWithoutPhoneNumber
 		Switch; Only return contacts that have a phone or mobile number
 
-	.PARAMETER IncludeNonMailboxContacts
-		Switch; Also include directory users that don't have an actual mailbox
+	.PARAMETER ExcludeSharedMailboxContacts
+		Switch; Excludes contacts that are a shared mailbox, or a mailbox without a liscense
+
+	.PARAMETER IncludeNonUserContacts
+		Switch; Also return contacts that aren't users/mailboxes in your directory. These contacts must still have an email address.
 	
 	.EXAMPLE
 		PS C:\> Get-GALContacts -ConnectionUri 'https://outlook.office365.com/powershell-liveid/' -Credentials $Credentials
@@ -30,36 +33,48 @@ param (
 
 	[Parameter(Position = 2, Mandatory = $false)]
 	[bool]
-	$RequirePhoneNumber,
+	$ExcludeContactsWithoutPhoneNumber,
 
 	[Parameter(Position = 3, Mandatory = $false)]
 	[bool]
-	$IncludeNonMailboxContacts
+	$ExcludeSharedMailboxContacts,
+
+	[Parameter(Position = 4, Mandatory = $false)]
+	[bool]
+	$IncludeNonUserContacts
 )
 process {
-	try {
+	#try {
 		# Connect to Office 365 Exchange Server using a Remote Session
 	$Session = New-PSSession -ConfigurationName Microsoft.Exchange -ConnectionUri $ConnectionUri -Credential $Credentials -Authentication Basic -AllowRedirection
 	Import-PSSession $Session -DisableNameChecking -AllowClobber
 	
 		# Import Global Address List into Powershell from Office 365 Exchange as an array
-		$ContactList = Get-User -ResultSize unlimited | Where-Object {$null -ne $_.WindowsEmailAddress}
+		$ContactList = Get-User -ResultSize unlimited 
 
-		# If the IncludeNonMailboxContacts switch is enabled, also include contacts that don't have a mailbox in your directory.
-		if ($IncludeNonMailboxContacts) {
-			$ContactList = $ContactList | Select-Object DisplayName,FirstName,LastName,Title,Company,Department,WindowsEmailAddress,Phone,MobilePhone
-		} else {
+		# If the ExcludeSharedMailboxContacts switch is enabled, exclude contacts that are a shared mailbox or mailbox with no liscense
+		if ($ExcludeSharedMailboxContacts) {
 			$DirectoryList = $(Get-Mailbox -ResultSize unlimited | Where-Object {$_.HiddenFromAddressListsEnabled -Match "False"})
 			$EmailAddressList = $DirectoryList.PrimarySMTPAddress
 			$ContactList = $ContactList | Select-Object DisplayName,FirstName,LastName,Title,Company,Department,WindowsEmailAddress,Phone,MobilePhone | Where-Object {$EmailAddressList.Contains($_.WindowsEmailAddress)}
+		} else {
+			$ContactList = $ContactList | Select-Object DisplayName,FirstName,LastName,Title,Company,Department,WindowsEmailAddress,Phone,MobilePhone
 		}
-		if ($RequirePhoneNumber) {
+		
+		# If the IncludeNonUserContacts switch is enabled, also include contacts that aren't actual users in the directory
+		if ($IncludeNonUserContacts) {
+			$ContactList += Get-Contact | Select-Object DisplayName,FirstName,LastName,Title,Company,Department,WindowsEmailAddress,Phone,MobilePhone
+		}
+
+		# If the ExcludeContactsWithoutPhoneNumber switch is enabled, exclude contacts that don't have a phone or mobile number
+		if ($ExcludeContactsWithoutPhoneNumber) {
 			$ContactList = $ContactList | Where-Object {$_.Phone -or $_.MobilePhone}
 		}
 	Remove-PSSession $Session
-	return $ContactList
-	} catch {
+	# Only return contacts with email addresses
+	return $ContactList | Where-Object {$null -ne $_.WindowsEmailAddress -and "" -ne $_.WindowsEmailAddress}
+	#} catch {
 		Write-Log -Level "FATAL" -Message "Failed to fetch Global Address List Contacts from Office 365 Directory" -exception $_.Exception.Message
-	}
+	#}
 }
 }